Demo file for next week
This is the base code for the demonstration for the TA's next week. We'll drive the robot in a box-ish and find distance, then lift elevator. Distance needs to be added (@chandler, obstacle avoidance) and robot needs to be built (Rachel) and wired (Nick) and code tweaked as needed (Madeline).